+++ src/main/java/com/minecolonies/coremod/network/messages/server/colony/building/enchanter/EnchanterWorkerSetMessage.java
@@ -1,6 +1,8 @@
 package com.minecolonies.coremod.network.messages.server.colony.building.enchanter;
 
 import com.minecolonies.api.colony.IColony;
+import com.minecolonies.api.colony.buildings.views.IBuildingView;
+import com.minecolonies.coremod.colony.buildings.modules.EnchanterStationsModule;
 import com.minecolonies.coremod.colony.buildings.workerbuildings.BuildingEnchanter;
 import com.minecolonies.coremod.network.messages.server.AbstractBuildingServerMessage;
 import net.minecraft.network.PacketBuffer;
@@ -38,7 +40,7 @@ public class EnchanterWorkerSetMessage extends AbstractBuildingServerMessage<Bui
      * @param worker   the worker to add/remove.
      * @param add      true if add, else false
      */
-    public EnchanterWorkerSetMessage(@NotNull final BuildingEnchanter.View building, final BlockPos worker, final boolean add)
+    public EnchanterWorkerSetMessage(@NotNull final IBuildingView building, final BlockPos worker, final boolean add)
     {
         super(building);
         this.worker = worker;
@@ -48,7 +50,6 @@ public class EnchanterWorkerSetMessage extends AbstractBuildingServerMessage<Bui
     @Override
     public void fromBytesOverride(@NotNull final PacketBuffer buf)
     {
-
         worker = buf.readBlockPos();
         add = buf.readBoolean();
     }
@@ -56,22 +57,20 @@ public class EnchanterWorkerSetMessage extends AbstractBuildingServerMessage<Bui
     @Override
     public void toBytesOverride(@NotNull final PacketBuffer buf)
     {
-
         buf.writeBlockPos(worker);
         buf.writeBoolean(add);
     }
 
     @Override
-    protected void onExecute(
-      final NetworkEvent.Context ctxIn, final boolean isLogicalServer, final IColony colony, final BuildingEnchanter building)
+    protected void onExecute(final NetworkEvent.Context ctxIn, final boolean isLogicalServer, final IColony colony, final BuildingEnchanter building)
     {
         if (add)
         {
-            building.addWorker(worker);
+            building.getFirstModuleOccurance(EnchanterStationsModule.class).addWorker(worker);
         }
         else
         {
-            building.removeWorker(worker);
+            building.getFirstModuleOccurance(EnchanterStationsModule.class).removeWorker(worker);
         }
     }
 }
